+++ app/(routes)/product/[id]/_components/index.ts
