+++ app/_components/index.ts
@@ -2,8 +2,7 @@ export { default as FilterChip } from './layout/FilterChip/FilterChip';
 export { default as FilterChipsBox } from './layout/FilterChip/FilterChipsBox';
 
 export { default as SideNavigationBar } from './layout/SideNavigationBar/SideNavigationBar';
-export { default as SNBFilterContainer } from './layout/SideNavigationBar/SNBFilterContainer';
-export { default as SNBFilterList } from './layout/SideNavigationBar/SNBFilterList';
+export { default as SNBList } from './layout/SideNavigationBar/SNBList';
 
 export { default as FloatingUpButton } from './layout/FloatingUpButton';
 export { default as GlobalNavigationBar } from './layout/GlobalNavigationBar';
