+++ app/_components/layout/SideNavigationBar/SNBList.tsx
@@ -1,7 +1,7 @@
 import { Grid, ListItem, Typography } from '@mui/material';
 import clsx from 'clsx';
 
-interface FilterListProps {
+interface ISNBListProps {
   data: string[];
   selectFilter: string;
   onSelectFilter: (name: string) => void;
@@ -10,11 +10,7 @@ interface FilterListProps {
 /**
  * 사이드 네비게이션 바 > 필터 리스트 컴포넌트
  */
-function SNBFilterList({
-  data,
-  selectFilter,
-  onSelectFilter,
-}: FilterListProps) {
+function SNBList({ data, selectFilter, onSelectFilter }: ISNBListProps) {
   return (
     <Grid item xs={6}>
       {data.map((name: string) => (
@@ -34,4 +30,4 @@ function SNBFilterList({
   );
 }
 
-export default SNBFilterList;
+export default SNBList;
